Centralizacion de conexion y carga de DGV

+++ App-Certificado/ProyectoFinal/Dominio/Articulos.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Data.Sql;
+using System.Data.SqlTypes;
 
 namespace Dominio //Cada parte de la tabla que necesito recrear en el codigo
 {
@@ -16,7 +18,7 @@ namespace Dominio //Cada parte de la tabla que necesito recrear en el codigo
         public int IdMarca { get; set; }
         public int IdCategoria { get; set; }
         public string UrlImagen { get; set; }
-        public float Precio { get; set; }
+        public decimal Precio { get; set; }
 
         //METODOS
